wire up workers too
