+++ koku/sources/management/commands/sources_listener.py
@@ -4,42 +4,43 @@
 #
 """Sources listener entry point."""
 import logging
-import threading
 import time
-from http.server import HTTPServer
 
 from django.core.management.base import BaseCommand
 
 from koku.database import check_migrations
-from koku.env import ENVIRONMENT
+from koku.probe_server import ProbeResponse
+from koku.probe_server import ProbeServer
+from koku.probe_server import start_probe_server
+from sources.api.status import check_kafka_connection
+from sources.api.status import check_sources_connection
 from sources.kafka_listener import initialize_sources_integration
-from sources.management.commands.probe_server import SourcesProbeServer
 
 
 LOG = logging.getLogger(__name__)
-CLOWDER_PORT = 8080
-if ENVIRONMENT.bool("CLOWDER_ENABLED", default=False):
-    from app_common_python import LoadedConfig
-
-    CLOWDER_PORT = LoadedConfig.publicPort
 
 
-def start_probe_server():
-    """Start the probe server."""
-    httpd = HTTPServer(("0.0.0.0", CLOWDER_PORT), SourcesProbeServer)
+class SourcesProbeServer(ProbeServer):
+    """HTTP server for liveness/readiness probes."""
 
-    def start_server():
-        """Start a simple webserver serving path on port"""
-        httpd.RequestHandlerClass.ready = False
-        httpd.serve_forever()
-
-    LOG.info("starting liveness/readiness probe server")
-    daemon = threading.Thread(name="probe_server", target=start_server)
-    daemon.setDaemon(True)  # Set as a daemon so it will be killed once the main thread is dead.
-    daemon.start()
-    LOG.info("liveness/readiness probe server started")
-
-    return httpd
+    def readiness_check(self):
+        """Set the readiness check response."""
+        status = 424
+        msg = "not ready"
+        if self.ready:
+            if not check_kafka_connection():
+                response = ProbeResponse(status, "kafka connection error")
+                self._write_response(response)
+                self.logger.info(response.json)
+                return
+            if not check_sources_connection():
+                response = ProbeResponse(status, "sources-api not ready")
+                self._write_response(response)
+                self.logger.info(response.json)
+                return
+            status = 200
+            msg = "ok"
+        self._write_response(ProbeResponse(status, msg))
 
 
 class Command(BaseCommand):
@@ -48,7 +49,7 @@ class Command(BaseCommand):
     help = "Starts koku-sources-kafka-listener"
 
     def handle(self, *args, **kwargs):
-        httpd = start_probe_server()
+        httpd = start_probe_server(SourcesProbeServer)
 
         timeout = 5
         while not check_migrations():
